backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
pymysql
psycopg2-binary